deps: update to ES Java client 8.15.5 in Optimize

+++ optimize/util/optimize-commons/src/main/java/io/camunda/optimize/service/util/configuration/ElasticSearchConfiguration.java
@@ -93,14 +93,14 @@ public class ElasticSearchConfiguration {
   }
 
   @JsonIgnore
-  public Integer getNestedDocumentsLimit() {
-    final Integer values = settings.getIndex().getNestedDocumentsLimit();
+  public Long getNestedDocumentsLimit() {
+    final Long values = settings.getIndex().getNestedDocumentsLimit();
     ensureGreaterThanZero(values);
     return values;
   }
 
   @JsonIgnore
-  public void setNestedDocumentsLimit(final int nestedDocumentLimit) {
+  public void setNestedDocumentsLimit(final long nestedDocumentLimit) {
     settings.getIndex().setNestedDocumentsLimit(nestedDocumentLimit);
   }
